feat(lbac-2357): ajout autocompletion pour la creation d entreprises par les cfa, et interdiction creation de siret CFA
