+++ server/src/http/controllers/etablissementRecruteur.controller.ts
@@ -1,5 +1,5 @@
 import { badRequest, forbidden, internal, notFound } from "@hapi/boom"
-import { assertUnreachable, toPublicUser, TrafficType, zRoutes } from "shared"
+import { assertUnreachable, IEntreprise, toPublicUser, TrafficType, zRoutes } from "shared"
 import { CFA, ENTREPRISE } from "shared/constants"
 import { BusinessErrorCodes } from "shared/constants/errorCodes"
 import { OPCOS_LABEL, RECRUITER_STATUS } from "shared/constants/recruteur"
@@ -74,25 +74,24 @@ export default (server: Server) => {
         throw badRequest("Le numéro siret est obligatoire.")
       }
 
-      const cfaVerification = await validateCreationEntrepriseFromCfa({ siret, cfa_delegated_siret })
+      let entrepriseOpt: IEntreprise | null = null
+      if (skipUpdate) {
+        entrepriseOpt = await getDbCollection("entreprises").findOne({ siret })
+      }
+      if (!entrepriseOpt) {
+        const siretResponse = await getEntrepriseDataFromSiret({ siret, type: cfa_delegated_siret ? CFA : ENTREPRISE })
+        entrepriseOpt = await upsertEntrepriseData(siret, "création de compte entreprise", siretResponse, false)
+        if ("error" in siretResponse) {
+          throw badRequest(siretResponse.message, siretResponse)
+        }
+      }
+
+      const cfaVerification = await validateCreationEntrepriseFromCfa({ siret, cfa_delegated_siret, nafCode: entrepriseOpt.naf_code ?? undefined })
       if (cfaVerification) {
         throw badRequest(cfaVerification.message)
       }
 
-      if (skipUpdate) {
-        const entrepriseOpt = await getDbCollection("entreprises").findOne({ siret })
-        if (entrepriseOpt) {
-          return res.status(200).send(entrepriseOpt)
-        }
-      }
-      const siretResponse = await getEntrepriseDataFromSiret({ siret, type: cfa_delegated_siret ? CFA : ENTREPRISE })
-      const entreprise = await upsertEntrepriseData(siret, "création de compte entreprise", siretResponse, false)
-
-      if ("error" in siretResponse) {
-        throw badRequest(siretResponse.message, siretResponse)
-      } else {
-        return res.status(200).send(entreprise)
-      }
+      return res.status(200).send(entrepriseOpt)
     }
   )
